Cambios en el modelo
Se agregaron cambios al modelo, se hizo in script provisional con los
permisos que llevara el usuario

+++ src/com/puntoVenta/Producto.java
@@ -12,7 +12,7 @@ package com.puntoVenta;
 public class Producto {
     private int idProducto;
     private String nombreProducto;
-    private int precio;
+    private float precio;
 
     public Producto() {
     }
